+++ lang/en/mumie.php
@@ -110,6 +110,7 @@ $string['mumie_form_server_btn_submit'] = 'Submit';
 $string['mumie_form_server_btn_cancel'] = 'Cancel';
 $string['mumie_url_prefix'] = 'MUMIE URL Prefix';
 $string['mumie_url_prefix_help'] = 'Specify the MUMIE URL prefix  <br><br> e.g. <b>https://www.ombplus.de/ombplus</b> <br><br> There can only be a single configuration for any URL prefix.';
+$string['mumie_form_cant_change_isgraded'] = 'You cannot switch from graded to ungraded problems. Please create a new MUMIE Task instead.';
 
 // Used in duedate form.
 $string['mumie_duedate_form'] = 'Grant a due date extension';