+++ changelog.md
@@ -2,6 +2,13 @@
 
 All important changes to this plugin will be documented in this file.
 
+## [v1.4.4] - [TODO]
+### Added
+- Select ungraded problems in problem selector
+
+### Changed
+- Removed "use entire course" button. You can create a MUMIE Task linking to the course overview page by selecting the respective problem in problem selector. It's listed under **Type -> Overview**.
+
 ## [v1.4.3] - 2021-06-25
 ### Fixed
 - Fixed an issue on the MUMIE grading page where sorting the student list by either *Due date extension* or *Submissions* caused an error